# docs/api.yaml
openapi: 3.0.3
info:
  title: EnergyMind API
  version: 0.1.0
  description: API contract for the EnergyMind application.
servers:
  - url: http://localhost:8000
paths:
  /health:
    get:
      summary: Проверка доступности API
      responses:
        "200":
          description: API доступен
  /api/premises:
    get:
      summary: Получить помещения объекта
      parameters:
        - $ref: "#/components/parameters/TenantId"
      responses:
        "200":
          description: Список помещений
    post:
      summary: Создать или обновить помещение
      parameters:
        - $ref: "#/components/parameters/TenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Premise"
      responses:
        "201":
          description: Помещение сохранено
  /api/datasets:
    post:
      summary: Загрузить CSV с почасовым потреблением
      parameters:
        - $ref: "#/components/parameters/TenantId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Набор данных загружен
    delete:
      summary: Очистить набор данных и результат анализа объекта
      parameters:
        - $ref: "#/components/parameters/TenantId"
      responses:
        "204":
          description: Данные объекта очищены
  /api/analyses:
    post:
      summary: Найти аномалии и построить прогноз на семь дней
      parameters:
        - $ref: "#/components/parameters/TenantId"
      responses:
        "200":
          description: Результат анализа
  /api/tasks:
    get:
      summary: Получить задачи объекта
      parameters:
        - $ref: "#/components/parameters/TenantId"
      responses:
        "200":
          description: Список задач
    post:
      summary: Подтвердить рекомендацию созданием задачи
      parameters:
        - $ref: "#/components/parameters/TenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [recommendationId, title]
              properties:
                recommendationId:
                  type: string
                title:
                  type: string
                assignee:
                  type: string
      responses:
        "201":
          description: Задача создана
  /buildings/main:
    get:
      summary: Получить модель здания для интерактивного плана
      parameters:
        - $ref: "#/components/parameters/TenantId"
      responses:
        "200":
          description: Здание и помещения с текущими статусами
  /metrics/dashboard:
    get:
      summary: Получить ключевые показатели дашборда
      parameters:
        - $ref: "#/components/parameters/TenantId"
      responses:
        "200":
          description: Потребление, число аномалий и потенциальная экономия
  /anomalies:
    get:
      summary: Получить аномалии в формате дашборда
      parameters:
        - $ref: "#/components/parameters/TenantId"
      responses:
        "200":
          description: Список аномалий
  /rooms/{roomId}/consumption:
    get:
      summary: Получить потребление помещения
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: Временной ряд фактического и базового потребления
  /rooms/{roomId}/devices:
    get:
      summary: Получить устройства помещения
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: Список устройств и их потребление
  /rooms/{roomId}/recommendation:
    get:
      summary: Получить рекомендацию для помещения
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: Рекомендация либо null при её отсутствии
  /analysis/run:
    post:
      summary: Запустить анализ для дашборда
      parameters:
        - $ref: "#/components/parameters/TenantId"
      responses:
        "200":
          description: Краткий результат анализа
  /tasks:
    post:
      summary: Создать задачу из карточки помещения
      parameters:
        - $ref: "#/components/parameters/TenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roomId, title, description]
              properties:
                roomId:
                  type: string
                title:
                  type: string
                description:
                  type: string
      responses:
        "201":
          description: Задача создана
components:
  parameters:
    TenantId:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
      description: Идентификатор объекта. Данные изолированы по этому значению.
    RoomId:
      name: roomId
      in: path
      required: true
      schema:
        type: string
  schemas:
    Premise:
      type: object
      required: [id, type, areaM2]
      properties:
        id:
          type: string
          example: A-203
        type:
          type: string
          example: classroom
        areaM2:
          type: number
          example: 48
        workingHours:
          type: string
          example: 09-19
        hasAirConditioning:
          type: boolean
        hasLighting:
          type: boolean